Refactor logout functionality and improve UI components
- Updated logout process in AuthCubit to clear all stored data and reset application state.
- Enhanced navigation logic in MainScreenV2 and SettingPage to clear the navigation stack upon logout.
- Improved layout in ResultScreen to provide a more structured display of results.
- Adjusted sizing in ChatScreen for better responsiveness.
- Ensured proper data type handling in Longevity widget for score calculations.

+++ lib/widgets/longevity.dart
@@ -42,11 +42,11 @@ class _LongevityState extends State<Longevity> {
         final data = response.data["scores"];
         print(data);
         setState(() {
-          dietScore = data['diet'] ?? 0;
-          mindScore = data['lifestyle'] ?? 0;
-          activityScore = data['activity'] ?? 0;
-          supplementScore = data['supplement'] ?? 0;
-          lifestyle = data['lifestyle'] ?? 0;
+          dietScore = data['diet'].toDouble() ?? 0;
+          mindScore = data['lifestyle'].toDouble() ?? 0;
+          activityScore = data['activity'].toDouble() ?? 0;
+          supplementScore = data['supplement'].toDouble() ?? 0;
+          lifestyle = data['lifestyle'].toDouble() ?? 0;
           isLoading = false;
         });
       } else {
